Almost done. Missing a few commands

+++ include/Utils/FileDeleter.hpp
@@ -5,6 +5,6 @@
 class FileDeleter : public FileHandler
 {
 public:
-    FileDeleter(const std::string user, const unsigned int clientId, Mutex* mutex, const std::string filename);
+    FileDeleter(const std::string user, const unsigned long long clientId, Mutex* mutex, const std::string filename);
     ~FileDeleter(){};
 };
