+++ src/Utils/FileDeleter.cpp
@@ -1,3 +1,3 @@
 #include "Utils/FileDeleter.hpp"
 
-FileDeleter::FileDeleter(const std::string user, const unsigned int clientId, Mutex* mutex, const std::string filename) : FileHandler(user, clientId, mutex, filename, NULL, FileHandlerMode::DELETE){}
+FileDeleter::FileDeleter(const std::string user, const unsigned long long clientId, Mutex* mutex, const std::string filename) : FileHandler(user, clientId, mutex, filename, NULL, FileHandlerMode::DELETE){}
